+++ generate-new/vk/VK/SPIRVElements.hs
@@ -4,6 +4,9 @@ module VK.SPIRVElements
   ( renderSPIRVElements
   ) where
 
+import           CType                          ( CType(TypeName) )
+import           Data.Bits
+import           Data.Foldable
 import qualified Data.Text                     as T
 import           Data.Text.Prettyprint.Doc
 import           Data.Vector                    ( Vector )
@@ -12,40 +15,54 @@ import           Data.Version                   ( Version(versionBranch)
                                                 , makeVersion
                                                 )
 import           Error
+import           Haskell                        ( HName(..)
+                                                , allTypeNames
+                                                , renderType
+                                                , (~>)
+                                                )
+import           Language.Haskell.TH            ( Type(ConT)
+                                                , mkName
+                                                )
 import           Polysemy.Input
 import           Relude
 import           Render.Element
+import           Render.SpecInfo                ( HasSpecInfo )
+import           Render.Type                    ( cToHsType )
+import           Render.Type.Preserve           ( Preserve(DoNotPreserve) )
 import           Spec.Types
 import           Text.InterpolatedString.Perl6.Unindented
 import           VkModulePrefix
 
 renderSPIRVElements
-  :: (HasErr r, HasRenderParams r)
+  :: (HasErr r, HasRenderParams r, HasSpecInfo r)
   => Vector SPIRVExtension
   -> Vector SPIRVCapability
   -> Sem r RenderElement
-renderSPIRVElements exts caps = do
-  genRe "SPIR-V stuff" $ do
-    tellExplicitModule (vulkanModule ["SPIRVInfo"])
-    tellCanFormat
-    bespokeStuff
-    renderExts exts
-    renderCaps caps
+renderSPIRVElements exts caps = genRe "SPIR-V stuff" $ do
+  tellExplicitModule (vulkanModule ["SPIRVRequirements"])
+  tellCanFormat
+  bespokeStuff
+  renderExts exts
+  renderCaps caps
 
 renderExts
-  :: (HasRenderElem r, HasRenderParams r) => Vector SPIRVExtension -> Sem r ()
+  :: (HasRenderElem r, HasRenderParams r, HasErr r, HasSpecInfo r)
+  => Vector SPIRVExtension
+  -> Sem r ()
 renderExts = renderSPIRVThing "spirvExtensionRequirements"
                               spirvExtensionName
                               spirvExtensionReqs
 
 renderCaps
-  :: (HasRenderElem r, HasRenderParams r) => Vector SPIRVCapability -> Sem r ()
+  :: (HasRenderElem r, HasRenderParams r, HasErr r, HasSpecInfo r)
+  => Vector SPIRVCapability
+  -> Sem r ()
 renderCaps = renderSPIRVThing "spirvCapabilityRequirements"
                               spirvCapabilityName
                               spirvCapabilityReqs
 
 renderSPIRVThing
-  :: (HasRenderElem r, HasRenderParams r)
+  :: (HasRenderElem r, HasRenderParams r, HasErr r, HasSpecInfo r)
   => Text
   -> (a -> Text)
   -> (a -> Vector SPIRVRequirement)
@@ -57,38 +74,111 @@ renderSPIRVThing funName name reqs xs = do
         pure $ viaShow (name x) <+> "->" <+> list reqs'
   cases <- (<> ["_ -> []"]) <$> traverse case' (V.toList xs)
   tellImport ''ByteString
+  tellImport (mkName "Vulkan.Requirements.Requirement")
+  tellExport (ETerm (TermName funName))
   tellDoc $ vsep
-    [ (pretty funName <+> ":: ByteString -> [SPIRVRequirement]")
-    , (pretty funName <+> "= \\case" <> line <> indent 2 (vsep cases))
+    [ pretty funName <+> ":: ByteString -> [Requirement]"
+    , pretty funName <+> "= \\case" <> line <> indent 2 (vsep cases)
     ]
 
 renderReq
-  :: (HasRenderParams r, HasRenderElem r) => SPIRVRequirement -> Sem r (Doc ())
+  :: (HasRenderParams r, HasRenderElem r, HasErr r, HasSpecInfo r)
+  => SPIRVRequirement
+  -> Sem r (Doc ())
 renderReq = \case
-  SPIRVReqVersion v -> do
-    tellImport 'makeVersion
-    pure $ "Version $ makeVersion" <+> viaShow (versionBranch v)
+  SPIRVReqVersion   v -> versionReq v
   SPIRVReqExtension p -> do
-    p' <- extensionNamePattern p
-    pure $ "Extension" <+> p'
+    extensionNamePattern p >>= \case
+      Left  v  -> versionReq v
+      Right p' -> do
+        tellImportWithAll (mkName "Vulkan.Requirements.Requirement")
+        pure $ "RequireExtension" <+> p'
   SPIRVReqFeature s f rs -> do
-    rs' <- traverse extensionNamePattern (V.toList rs)
+    RenderParams {..} <- input
+    tellImportWithAll (mkName "Vulkan.Requirements.Requirement")
     tellLanguageExtension (LanguageExtension "OverloadedLists")
-    pure $ "Feature" <+> viaShow s <+> viaShow f <+> list rs'
+    sTy <- cToHsType DoNotPreserve (TypeName s)
+    -- TODO: this is pretty lazy, import the accessors properly
+    traverse_ tellImportWithAll (allTypeNames sTy)
+    checkTDoc    <- renderType (sTy ~> ConT ''Bool)
+    sTyDoc       <- renderType sTy
+    otherSetters <- minVersionAndExtensionsSetters rs
+    let featureMemberName = mkMemberName f
+    let xs =
+          [ "featureName =" <+> viaShow f
+            , "checkFeature =" <+> pretty featureMemberName <+> "::" <+> checkTDoc
+            , "enableFeature ="
+            <+> "\\f ->"
+            <+> "f"
+            <>  braces (pretty featureMemberName <+> "= True")
+            <+> "::"
+            <+> sTyDoc
+            ]
+            <> otherSetters
+    pure $ "RequireFeature" <> encloseSep "{" "}" ", " xs
   SPIRVReqProperty p m v rs -> do
-    rs' <- traverse extensionNamePattern (V.toList rs)
+    RenderParams {..} <- input
+    tellImportWithAll (mkName "Vulkan.Requirements.Requirement")
     tellLanguageExtension (LanguageExtension "OverloadedLists")
-    pure $ "Property" <+> viaShow p <+> viaShow m <+> viaShow v <+> list rs'
+    -- TODO: this is pretty lazy, import the accessors properly
+    sTy <- cToHsType DoNotPreserve (TypeName p)
+    traverse_ tellImportWithAll (allTypeNames sTy)
+    sTyDoc <- renderType sTy
+    let propertyMemberName = mkMemberName m
+        propertyValueName  = mkPatternName v
+    otherSetters <- minVersionAndExtensionsSetters rs
+    -- TODO, do this properly
+    checker      <- if
+      | v == "VK_TRUE"
+      -> pure $ "\\p ->" <+> pretty propertyMemberName <+> parens
+        ("p ::" <+> sTyDoc)
+      | "_BIT" `T.isInfixOf` unCName v
+      -> do
+        tellImport propertyValueName
+        pure
+          $   "\\p ->"
+          <+> pretty propertyValueName
+          <+> ".&&."
+          <+> pretty propertyMemberName
+          <+> parens ("p ::" <+> sTyDoc)
+      | otherwise
+      -> do
+        tellImport propertyValueName
+        pure
+          $   "\\p ->"
+          <+> pretty propertyValueName
+          <+> "=="
+          <+> pretty propertyMemberName
+          <+> parens ("p ::" <+> sTyDoc)
+    let xs =
+          ["propertyName =" <+> viaShow p, "checkProperty =" <+> checker]
+            <> otherSetters
+    pure $ "RequireProperty" <> encloseSep "{" "}" ", " xs
+
+minVersionAndExtensionsSetters
+  :: (HasRenderParams r, HasRenderElem r, HasErr r, HasSpecInfo r)
+  => Vector Text
+  -> Sem r [Doc ()]
+minVersionAndExtensionsSetters rs = do
+  (minVersions, extensionNames) <-
+    partitionEithers . V.toList <$> traverse extensionNamePattern rs
+  minVersionDoc <- case minVersions of
+    [] -> pure "Nothing"
+    vs -> ("Just $" <+>) <$> versionDoc (maximum vs)
+  pure
+    [ "requireMinVersion =" <+> minVersionDoc
+    , "requireExtensions =" <+> viaShow extensionNames
+    ]
 
 extensionNamePattern
   :: (HasRenderElem r, HasRenderParams r)
   => Text
-  -> Sem r (Doc ())
+  -> Sem r (Either Version (Doc ()))
 extensionNamePattern p = do
   RenderParams {..} <- input
-  if "VK_VERSION_" `T.isPrefixOf` p
-    then pure $ viaShow p
-    else do
+  case parseVersion p of
+    Just v  -> pure . Left $ v
+    Nothing -> do
       let nameName = case p of
             -- TODO: Handle these properly
             "VK_INTEL_shader_integer_functions2" ->
@@ -96,20 +186,42 @@ extensionNamePattern p = do
             _ -> CName $ T.toUpper p <> "_EXTENSION_NAME"
       let p' = mkPatternName nameName
       tellImport p'
-      pure $ pretty p'
+      pure . Right $ pretty p'
+
+versionReq
+  :: (HasRenderParams r, HasRenderElem r, HasErr r) => Version -> Sem r (Doc ())
+versionReq v = do
+  tellImportWithAll (mkName "Vulkan.Requirements.Requirement")
+  vDoc <- versionDoc v
+  pure $ "RequireVersion $" <+> vDoc
+
+versionDoc
+  :: (HasRenderParams r, HasRenderElem r, HasErr r) => Version -> Sem r (Doc ())
+versionDoc v = do
+  tellImport (ConName "MAKE_VERSION")
+  (ma, mi, pa) <- case versionBranch v of
+    [ma]         -> pure (ma, 0, 0)
+    [ma, mi]     -> pure (ma, mi, 0)
+    [ma, mi, pa] -> pure (ma, mi, pa)
+    []           -> throw "Version branch has no components"
+    _            -> throw "Version branch has more than three components"
+  pure $ "MAKE_VERSION" <+> hsep (viaShow <$> [ma, mi, pa])
+
+parseVersion :: Text -> Maybe Version
+parseVersion t = do
+  let p = "VK_VERSION_"
+  v <- if p `T.isPrefixOf` t then pure $ T.drop (T.length p) t else empty
+  let cs = T.split (== '_') v
+  is <- traverse (readMaybe . T.unpack) cs
+  pure $ makeVersion is
 
 bespokeStuff :: (HasRenderParams r, HasRenderElem r) => Sem r ()
 bespokeStuff = do
-  tellImport ''ByteString
-  tellImport ''Vector
-  tellImport ''Version
+  tellImport ''Bits
+  tellImport '(.&.)
+  tellImport 'zeroBits
   tellDoc [qqi|
-    data SPIRVRequirement
-      = Version Version
-      | Extension ByteString
-      | Feature ByteString ByteString (Vector ByteString)
-      -- ^ Struct, feature, requires
-      | Property ByteString ByteString ByteString (Vector ByteString)
-      -- ^ Property, member, value, requires
-      deriving(Show)
+    -- | Check if the intersection of bits is non-zero
+    (.&&.) :: Bits a => a -> a -> Bool
+    x .&&. y = (x .&. y) /= zeroBits
   |]