+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## WIP
 
-- Expose SPIR-V Extensions and Capabilities in `Vulkan.SPIRVInfo`
+- Expose SPIR-V Extensions and Capabilities in `Vulkan.SPIRVRequirements`
 
 ## [3.7] - 2020-11-24
 - Bump API version to v1.2.162